requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
python-dotenv
passlib
PyJWT
